requirements.txt
opencv-python==4.9.0.80
numpy==1.26.4
pytesseract==0.3.10
Pillow==10.2.0
requests==2.32.2
pywin32==306
obs-websocket-py==0.6.0